requirements.txt
absl-py==1.4.0
aiobotocore==2.5.2
aiohttp==3.8.5
aiohttp-retry==2.8.3
aioitertools==0.11.0
aiosignal==1.3.1
altair==5.1.1
amqp==5.1.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.2
asyncssh==2.13.2
atpublic==4.0
attrs==23.1.0
billiard==4.1.0
blinker==1.6.2
boto3==1.26.161
botocore==1.29.161
cachetools==5.3.1
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.8
cryptography==41.0.3
dataclasses-json==0.5.14
dictdiffer==0.9.0
diskcache==5.6.1
distlib==0.3.7
distro==1.8.0
dnspython==2.3.0
dpath==2.1.6
dulwich==0.21.5
dvc==3.11.1
dvc-data==2.10.1
dvc-http==2.30.2
dvc-objects==0.24.1
dvc-render==0.5.3
dvc-s3==2.23.0
dvc-studio-client==0.11.0
dvc-task==0.3.0
filelock==3.12.2
Flask==2.2.2
Flask-PyMongo==2.3.0
flatbuffers==23.5.26
flatten-dict==0.4.2
flufl.lock==7.1.1
frozenlist==1.4.0
fsspec==2023.6.0
funcy==2.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grandalf==0.8
greenlet==2.0.2
grpcio==1.57.0
h5py==3.9.0
huggingface-hub==0.16.4
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
iterative-telemetry==0.0.8
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keras==2.13.1
kombu==5.3.1
langchain==0.0.281
langsmith==0.0.33
libclang==16.0.6
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.20.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nanotime==0.5.2
networkx==3.1
numexpr==2.8.5
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.3.0
openai==0.28.0
opencv-python==4.8.0.74
opt-einsum==3.3.0
orjson==3.9.2
packaging==23.1
pandas==2.1.0
pathspec==0.11.2
Pillow==9.5.0
platformdirs==3.10.0
pluggy==1.2.0
prompt-toolkit==3.0.39
protobuf==4.24.2
psutil==5.9.5
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.0
pydot==1.4.2
pygit2==1.12.2
Pygments==2.15.1
pygtrie==2.5.0
pymongo==4.3.3
Pympler==1.0.1
pyparsing==3.1.1
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
s3fs==2023.6.0
s3transfer==0.6.1
safetensors==0.3.3
scmrepo==1.1.0
shortuuid==1.0.11
shtab==1.6.4
six==1.16.0
smmap==5.0.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.20
sqltrie==0.7.0
streamlit==1.26.0
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
tokenizers==0.13.3
toml==0.10.2
tomlkit==0.12.1
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
transformers==4.32.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.22.0
vine==5.0.0
virtualenv==20.24.2
voluptuous==0.13.1
wasabi==0.10.1
watchdog==3.0.0
wcwidth==0.2.6
websocket-client==1.6.1
Werkzeug==2.3.4
wrapt==1.15.0
yarl==1.9.2
zc.lockfile==3.0.post1
zipp==3.16.2
